+++ src/input/inputFileParser/QMInputParser.cpp
@@ -177,12 +177,14 @@ void QMInputParser::parseQMMethod(
         parseMaceQMMethod(method);
 
     else
-        throw InputFileException(std::format(
-            "Invalid qm_prog \"{}\" in input file.\n"
-            "Possible values are: dftbplus, ase_dftbplus, ase_xtb, pyscf, "
-            "turbomole, mace, mace_mp, mace_off",
-            lineElements[2]
-        ));
+        throw InputFileException(
+            std::format(
+                "Invalid qm_prog \"{}\" in input file.\n"
+                "Possible values are: dftbplus, ase_dftbplus, ase_xtb, pyscf, "
+                "turbomole, mace, mace_mp, mace_off",
+                lineElements[2]
+            )
+        );
 }
 
 /**
@@ -248,20 +250,7 @@ void QMInputParser::parseDispersion(
 {
     checkCommand(lineElements, lineNumber);
 
-    const auto dispersion = toLowerCopy(lineElements[2]);
-
-    if ("true" == dispersion || "yes" == dispersion || "on" == dispersion)
-        QMSettings::setUseDispersionCorrection(true);
-
-    else if ("false" == dispersion || "no" == dispersion || "off" == dispersion)
-        QMSettings::setUseDispersionCorrection(false);
-
-    else
-        throw InputFileException(std::format(
-            "Invalid dispersion \"{}\" in input file.\n"
-            "Possible values are: true, yes, on, false, no, off",
-            lineElements[2]
-        ));
+    QMSettings::setUseDispersionCorrection(keywordToBool(lineElements));
 }
 
 /**
@@ -319,13 +308,15 @@ void QMInputParser::parseMaceModelSize(
         QMSettings::setMaceModelSize(CUSTOM);
 
     else
-        throw InputFileException(std::format(
-            "Invalid mace_model_size \"{}\" in input file.\n"
-            "Possible values are: small, medium, large, small-0b,\n"
-            "medium-0b, small-0b2, medium-0b2, large-0b2, medium-0b3,\n"
-            "medium-mpa-0, medium-omat-0, custom",
-            lineElements[2]
-        ));
+        throw InputFileException(
+            std::format(
+                "Invalid mace_model_size \"{}\" in input file.\n"
+                "Possible values are: small, medium, large, small-0b,\n"
+                "medium-0b, small-0b2, medium-0b2, large-0b2, medium-0b3,\n"
+                "medium-mpa-0, medium-omat-0, custom",
+                lineElements[2]
+            )
+        );
 }
 
 /**
@@ -367,17 +358,21 @@ void QMInputParser::parseMaceQMMethod(const std::string_view &model)
     }
 
     else if ("mace_anicc" == model || "mace_ani" == model)
-        throw InputFileException(std::format(
-            "The mace ani model is not supported in this version of PQ.\n"
-        ));
+        throw InputFileException(
+            std::format(
+                "The mace ani model is not supported in this version of PQ.\n"
+            )
+        );
 
     else
     {
-        throw InputFileException(std::format(
-            "Invalid mace type qm_method \"{}\" in input file.\n"
-            "Possible values are: mace (mace_mp), mace_off",
-            model
-        ));
+        throw InputFileException(
+            std::format(
+                "Invalid mace type qm_method \"{}\" in input file.\n"
+                "Possible values are: mace (mace_mp), mace_off",
+                model
+            )
+        );
     }
 
     QMSettings::setQMMethod(QMMethod::MACE);
@@ -416,11 +411,13 @@ void QMInputParser::parseSlakosType(
         QMSettings::setSlakosType(CUSTOM);
 
     else
-        throw InputFileException(std::format(
-            "Invalid slakos type \"{}\" in input file.\n"
-            "Possible values are: 3ob, matsci, custom",
-            lineElements[2]
-        ));
+        throw InputFileException(
+            std::format(
+                "Invalid slakos type \"{}\" in input file.\n"
+                "Possible values are: 3ob, matsci, custom",
+                lineElements[2]
+            )
+        );
 }
 
 /**
@@ -451,22 +448,7 @@ void QMInputParser::parseThirdOrder(
 {
     checkCommand(lineElements, lineNumber);
 
-    const auto third_order = toLowerCopy(lineElements[2]);
-
-    if ("on" == third_order || "yes" == third_order || "true" == third_order)
-        QMSettings::setUseThirdOrderDftb(true);
-
-    else if ("off" == third_order || "no" == third_order ||
-             "false" == third_order)
-        QMSettings::setUseThirdOrderDftb(false);
-
-    else
-        throw InputFileException(std::format(
-            "Invalid DFTB third_order request \"{}\" in input file.\n"
-            "Possible values are: on, yes, true, off, no, false",
-            lineElements[2]
-        ));
-
+    QMSettings::setUseThirdOrderDftb(keywordToBool(lineElements));
     QMSettings::setIsThirdOrderDftbSet(true);
 }
 
@@ -504,10 +486,12 @@ void QMInputParser::parseHubbardDerivs(
         }
         else
         {
-            throw InputFileException(std::format(
-                "Invalid hubbard_derivs format \"{}\" in input file.",
-                derivs
-            ));
+            throw InputFileException(
+                std::format(
+                    "Invalid hubbard_derivs format \"{}\" in input file.",
+                    derivs
+                )
+            );
         }
     }
 
@@ -543,9 +527,11 @@ void QMInputParser::parseXtbMethod(
         QMSettings::setXtbMethod(IPEA1);
 
     else
-        throw InputFileException(std::format(
-            "Invalid xTB method \"{}\" in input file.\n"
-            "Possible values are: GFN1-xTB, GFN2-xTB, IPEA1-xTB",
-            lineElements[2]
-        ));
+        throw InputFileException(
+            std::format(
+                "Invalid xTB method \"{}\" in input file.\n"
+                "Possible values are: GFN1-xTB, GFN2-xTB, IPEA1-xTB",
+                lineElements[2]
+            )
+        );
 }